refactor openfield post sorting to a cleaner layout
