--- P2_PostProcess/OpenField/Scores/open_field_firing_field_detection.py
@@ -1,58 +1,0 @@
-import pandas as pd
-from math import sqrt
-from skimage import data
-from skimage.feature import blob_dog, blob_log, blob_doh
-from skimage.color import rgb2gray
-import matplotlib.pyplot as plt
-
-
-def detect_firing_fields(spike_data, cluster, save_path):
-    image_gray = spike_data.firing_maps[cluster]
-    image = spike_data.firing_maps[cluster]
-    plt.imshow(image_gray)
-
-    blobs_log = blob_log(image_gray, max_sigma=30, num_sigma=10, threshold=.1)
-    # Compute radii in the 3rd column.
-    blobs_log[:, 2] = blobs_log[:, 2] * sqrt(2)
-
-    blobs_dog = blob_dog(image_gray, max_sigma=30, threshold=.1)
-    blobs_dog[:, 2] = blobs_dog[:, 2] * sqrt(2)
-
-    blobs_doh = blob_doh(image_gray, max_sigma=30, threshold=.01)
-
-    blobs_list = [blobs_log, blobs_dog, blobs_doh]
-    colors = ['yellow', 'lime', 'red']
-    titles = ['Laplacian of Gaussian', 'Difference of Gaussian',
-              'Determinant of Hessian']
-    sequence = zip(blobs_list, colors, titles)
-
-    fig, axes = plt.subplots(1, 3, figsize=(9, 3), sharex=True, sharey=True)
-    ax = axes.ravel()
-
-    for idx, (blobs, color, title) in enumerate(sequence):
-        ax[idx].set_title(title)
-        ax[idx].imshow(image, interpolation='nearest')
-        for blob in blobs:
-            y, x, r = blob
-            c = plt.Circle((x, y), r, color=color, linewidth=2, fill=False)
-            ax[idx].add_patch(c)
-        ax[idx].set_axis_off()
-
-    plt.tight_layout()
-    # plt.show()
-    plt.savefig(save_path + '/detection_result_gauss.png')
-    plt.close()
-
-
-
-#  this is here for testing
-def main():
-    print('-------------------------------------------------------------')
-
-    recording_folder = 'C:/Users/s1466507/Documents/Ephys/recordings/M5_2018-03-06_15-34-44_of'
-    spike_data_frame_path = recording_folder + '/MountainSort/DataFrames/spatial_firing.pkl'
-    spike_data = pd.read_pickle(spike_data_frame_path)
-    detect_firing_fields(spike_data, 5, recording_folder)
-
-if __name__ == '__main__':
-    main()
